Fix typedefs

+++ src/analyzer.hpp
@@ -13,8 +13,8 @@
 }
 
 struct HashFunc {
-    ull (* func) (char* item) = (ull (*) (char*)) poisons::UNINITIALIZED_PTR;
-    const char* func_name     = (const char*)     poisons::UNINITIALIZED_PTR;
+    unsigned long long (* func) (char* item) = (unsigned long long (*) (char*)) poisons::UNINITIALIZED_PTR;
+    const char*           func_name          = (const char*) poisons::UNINITIALIZED_PTR;
 };
 
 struct CollisionData {
@@ -25,13 +25,13 @@ struct CollisionData {
 // ==================== Prototypes ====================
 char* random_word(char* word, int len);
 
-ull constant_hash    (char* string);
-ull first_letter_hash(char* string);
-ull symbol_sum_hash  (char* string);
-ull string_len_hash  (char* string);
-ull roll_hash        (char* string);
-ull crc32_hash       (char* string);
-ull asm_crc32_hash   (char* string);
+unsigned long long constant_hash    (char* string);
+unsigned long long first_letter_hash(char* string);
+unsigned long long symbol_sum_hash  (char* string);
+unsigned long long string_len_hash  (char* string);
+unsigned long long roll_hash        (char* string);
+unsigned long long crc32_hash       (char* string);
+unsigned long long asm_crc32_hash   (char* string);
 
 CollisionData* get_collision_info (HashTable* table);
 int            test_table_speed   (const char* filename, int repeats=1, double fi_coef=0);
@@ -39,9 +39,9 @@ int            test_func_collision(const char* filename, const HashFunc* hash, c
 // ====================================================
 
 // ==================== CONSTS ====================
-const ull P_VALUE = 257;
+const unsigned long long P_VALUE = 257;
 
-const HashFunc ALL_HASH_FUNCS[] = {
+const HashFunc ALL_HASH_FUNCS[]  = {
     { constant_hash,     "Constant hash"     },
     { first_letter_hash, "First symbol hash" },
     { symbol_sum_hash,   "Symbols sum hash"  },
